+++ src/router/index.js
@@ -6,6 +6,8 @@ import vuxInput from '@/page/modules/explore/vuxInput'
 import vModel from '@/page/modules/vueApi/vModel'
 import input from '@/page/modules/vueApi/vModel/input'
 import setMap from '@/page/modules/es6/setMap'
+import proxy from '@/page/modules/es6/proxy'
+import replace from '@/page/modules/regExp/replace'
 Vue.use(Router)
 
 export default new Router({
@@ -14,6 +16,11 @@ export default new Router({
       path: '/',
       name: 'HelloWorld',
       component: HelloWorld
+    },
+    {
+      path: '/regexp/replace',
+      name: 'replace',
+      component: replace
     },
     {
       path: '/explore/callhook',
@@ -39,6 +46,11 @@ export default new Router({
       path: '/es6/setmap',
       name: 'setMap',
       component: setMap
+    },
+    {
+      path: '/es6/proxy',
+      name: 'proxy',
+      component: proxy
     }
   ]
 })
